Move modules classes to modules.py
